+++ app/javascript/vue/routes/endpoints/ObservationMatrix.js
@@ -14,6 +14,10 @@ export const ObservationMatrix = {
 
   columns: (id) => AjaxCall('get', `/${controller}/${id}/observation_matrix_columns.json`),
 
+  row: (params) => AjaxCall('get', `/${controller}/row.json`, { params }),
+
+  rowLabels: (id, params) => AjaxCall('get', `/${controller}/${id}/row_labels`, { params }),
+
   rows: (id, params) => AjaxCall('get', `/${controller}/${id}/observation_matrix_rows.json`, { params }),
 
   otusUseInMatrix: (params) => AjaxCall('get', `/${controller}/otus_used_in_matrices.json`, { params })