+++ app/javascript/vue/tasks/observation_matrices/matrix_row_coder/MatrixRowCoder/SingleObservationDescriptor/SingleObservationDescriptor.js
@@ -3,7 +3,6 @@ import { GetterNames } from '../../store/getters/getters'
 
 import summaryView from '../SummaryView/SummaryView.vue'
 import RadialAnnotator from 'components/radials/annotator/annotator'
-import singleObservationZoomedView from '../ZoomedView/SingleObservationZoomedView/SingleObservationZoomedView.vue'
 
 export default {
   computed: {
@@ -15,8 +14,6 @@ export default {
     const descriptorId = this.$props.descriptor.id
     const otuId = this.$store.state.taxonId
 
-    // this.$store.dispatch(ActionNames.RequestDescriptorDepictions, descriptorId);
-    // this.$store.dispatch(ActionNames.RequestDescriptorNotes, descriptorId);
     this.$store.dispatch(ActionNames.RequestObservations, {descriptorId, otuId})
       .then(_ => {
         const observations = this.$store.getters[GetterNames.GetObservationsFor](descriptorId)
@@ -25,13 +22,6 @@ export default {
         } else if (observations.length === 1) {
           const observation = observations[0]
           this.observation = observation
-
-          if (this.observation.id) {
-            // this.$store.dispatch(ActionNames.RequestObservationCitations, observation.id);
-            // this.$store.dispatch(ActionNames.RequestObservationConfidences, observation.id);
-            // this.$store.dispatch(ActionNames.RequestObservationDepictions, observation.id);
-            // this.$store.dispatch(ActionNames.RequestObservationNotes, observation.id);
-          }
         }
       })
   },
@@ -48,7 +38,6 @@ export default {
   props: ['descriptor'],
   components: {
     summaryView,
-    singleObservationZoomedView,
     RadialAnnotator
   }
 }