+++ app/javascript/vue/tasks/observation_matrices/matrix_row_coder/store/actions/requestConfidenceLevels.js
@@ -1,5 +1,0 @@
-import { MutationNames } from '../mutations/mutations'
-
-export default function ({ commit, state }) {
-  commit(MutationNames.SetConfidenceLevels, state.request.getConfidenceLevels())
-};