+++ app/javascript/vue/tasks/observation_matrices/matrix_row_coder/store/mutations/setDescriptorDepictions.js
@@ -1,6 +1,0 @@
-import transformDepictionForViewmodel from '../helpers/transformDepictionForViewmodel'
-
-export default function (state, args) {
-  const descriptor = state.descriptors.find(d => d.id === args.descriptorId)
-  descriptor.depictions = args.depictions.map(transformDepictionForViewmodel)
-};