+++ app/javascript/vue/tasks/observation_matrices/matrix_row_coder/store/store.js
@@ -13,7 +13,6 @@ export function makeInitialState (requestModule) {
     description: undefined,
     descriptors: [],
     observations: [],
-    confidenceLevels: null,
     units: {}
   }
 }